+++ Release-Notes.md
@@ -1,6 +1,13 @@
 
 # Xmlips Release Notes
 
+## v0.0.2
+
+Consistency of `Xml` parameters. They are all required and accept arrays and
+pipeline input. Empty and null arrays are allowed. Null items are not allowed.
+
+`Add-Xml`, `Find-Xml`: renamed `Name` -> `Tag`.
+
 ## v0.0.1
 
 The first preview.